requirements.txt
flask
jsonify
chromadb==0.5.5
pysqlite3-binary
requests